\name{15_deprecated}
\alias{use.theme}
\alias{plot2d.cell}
\alias{plot3d.bar}
\alias{plot2d.contour}
\alias{plot3d.surface}
\alias{plot2d.tricontour}
\alias{plot3d.trisurface}
\title{Deprecated Functions}
\description{Deprecated functions, please do not use.}
\usage{
use.theme (\dots)
plot2d.cell (\dots)
plot3d.bar (\dots)
plot2d.contour (\dots)
plot3d.surface (\dots)
plot2d.tricontour (\dots)
plot3d.trisurface (\dots)
}
\arguments{
\item{\dots}{.}
}
