\name{16_deprecated}
\alias{use.theme}
\alias{plot2d.contour}
\alias{plot3d.bar}
\alias{plot3d.surface}
\alias{litmus.rainbow.fit}
\title{Deprecated Functions}
\description{Deprecated functions, please do not use.}
\usage{
use.theme (\dots)
plot2d.contour (\dots)
plot3d.bar (\dots)
plot3d.surface (\dots)
litmus.rainbow.fit (\dots)
}
\arguments{
\item{\dots}{.}
}
