\name{51_simple_color_palettes}
\alias{gold.top.color}
\alias{blue.top.color}
\alias{green.top.color}
\alias{purple.top.color}
\alias{heat.top.color}
\alias{gold.side.color}
\alias{blue.side.color}
\alias{green.side.color}
\alias{purple.side.color}
\alias{heat.side.color}
\alias{gold.iso.colors}
\alias{blue.iso.colors}
\alias{green.iso.colors}
\alias{purple.iso.colors}
\alias{heat.iso.colors}
\title{Simple Color Palettes}
\description{Functions to create a single color, or vector of colors}
\usage{
gold.top.color ()
blue.top.color ()
green.top.color ()
purple.top.color ()
heat.top.color ()

gold.side.color ()
blue.side.color ()
green.side.color ()
purple.side.color ()
heat.side.color ()

gold.iso.colors ()
blue.iso.colors ()
green.iso.colors ()
purple.iso.colors ()
heat.iso.colors ()
}
\value{
The top and side functions, return a single (R color) string.\cr
The iso functions return a character (R color) vector.
}
\references{Refer to the vignette for an overview, references and better examples.}
\examples{
gold.top.color ()
gold.side.color ()

gold.iso.colors ()
}
