\name{71_deprecated}
\alias{plotf_cfield_3d}
\alias{barface}
\alias{litmus.rainbow.fit}
\title{Deprecated Functions}
\description{Deprecated functions, please do not use.}
\usage{
plotf_cfield_3d (\dots, reverse.z=FALSE)
barface ()
litmus.rainbow.fit (\dots)
}
\arguments{
\item{\dots, reverse.z}{.}
}
