\name{71_deprecated}
\alias{plotf_cfield_3d}
\alias{opt.top.color}
\alias{opt.side.color}
\alias{opt.litmus.fit}
\title{Deprecated Functions}
\description{Deprecated functions, please do not use.}
\usage{
plotf_cfield_3d (\dots, reverse.z=FALSE)

opt.top.color ()
opt.side.color ()
opt.litmus.fit ()

}
\arguments{
\item{\dots, reverse.z}{.}
}
